requirements.txt
# FastAPI et serveur
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de données PostgreSQL
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Validation de données
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Traitement de données
pandas==2.1.3
numpy==1.25.2
requests==2.31.0

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# Outils utilitaires
python-dotenv==1.0.0
logging-config==1.0.3
typer==0.9.0
rich==13.6.0

# Tests
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 